+++ tools/launchdev.sh
@@ -1,9 +1,41 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
+function usage() {
+    echo "Usage: $0 [start|stop|restart|startclean] [-r runner_count] [-g] [-x] [-c]" >&2
+}
+
+subcommand=$1; shift
 runner_count=1
-if [ "$#" -gt 1 ]; then
-    runner_count=${2}
-fi
+use_gunicorn=false
+copy_examples=false
+load_content=true
+
+while getopts ":r:gxc" o; do
+    case "${o}" in
+        r)
+            runner_count=${OPTARG}
+            ;;
+        g)
+            use_gunicorn=true
+            ;;
+        x)
+            copy_examples=true
+            ;;
+        c)
+            load_content=false
+            ;;
+        \?)
+            echo "Invalid option: -$OPTARG" >&2
+            usage
+            exit 2
+            ;;
+        :)
+            echo "Option -$OPTARG requires an argument." >&2
+            usage
+            exit 2
+            ;;
+    esac
+done
 
 function init(){
     ST2_BASE_DIR="/opt/stackstorm"
@@ -62,13 +94,21 @@ function st2start(){
         sudo mkdir -p $PACKS_BASE_DIR
     fi
 
+    VIRTUALENVS_DIR=$ST2_BASE_DIR/virtualenvs
+
     sudo mkdir -p $PACKS_BASE_DIR/default/sensors/
     sudo mkdir -p $PACKS_BASE_DIR/default/actions/
     sudo mkdir -p $PACKS_BASE_DIR/default/rules/
+    sudo mkdir -p $VIRTUALENVS_DIR
     sudo chown -R ${CURRENT_USER}:${CURRENT_USER_GROUP} $PACKS_BASE_DIR
+    sudo chown -R ${CURRENT_USER}:${CURRENT_USER_GROUP} $VIRTUALENVS_DIR
     cp -Rp ./contrib/core/ $PACKS_BASE_DIR
     cp -Rp ./contrib/packs/ $PACKS_BASE_DIR
-    cp -Rp ./contrib/examples $PACKS_BASE_DIR
+
+    if [ "$copy_examples" = true ]; then
+        echo "Copying examples from ./contrib/examples to $PACKS_BASE_DIR"
+        cp -Rp ./contrib/examples $PACKS_BASE_DIR
+    fi
 
     # activate virtualenv to set PYTHONPATH
     source ./virtualenv/bin/activate
@@ -83,9 +123,16 @@ function st2start(){
 
     # Run the st2 API server
     echo 'Starting screen session st2-api...'
-    screen -d -m -S st2-api ./virtualenv/bin/python \
-        ./st2api/bin/st2api \
-        --config-file $ST2_CONF
+    if [ "${use_gunicorn}" = true ]; then
+        echo '  using guicorn to run st2-api...'
+        export ST2_CONFIG_PATH=${ST2_CONF}
+        screen -d -m -S st2-api ./virtualenv/bin/gunicorn_pecan \
+            ./st2api/st2api/gunicorn_config.py -k eventlet
+    else
+        screen -d -m -S st2-api ./virtualenv/bin/python \
+            ./st2api/bin/st2api \
+            --config-file $ST2_CONF
+    fi
 
     # Start a screen for every runner
     echo 'Starting screen sessions for st2-actionrunner(s)...'
@@ -126,9 +173,16 @@ function st2start(){
 
     # Run the auth API server
     echo 'Starting screen session st2-auth...'
-    screen -d -m -S st2-auth ./virtualenv/bin/python \
+    if [ "${use_gunicorn}" = true ]; then
+        echo '  using guicorn to run st2-auth...'
+        export ST2_CONFIG_PATH=${ST2_CONF}
+        screen -d -m -S st2-auth ./virtualenv/bin/gunicorn_pecan \
+            ./st2auth/st2auth/gunicorn_config.py -k eventlet
+    else
+        screen -d -m -S st2-auth ./virtualenv/bin/python \
         ./st2auth/bin/st2auth \
         --config-file $ST2_CONF
+    fi
 
     if [ -n "$ST2_EXPORTER" ]; then
         EXPORTS_DIR=$(exportsdir)
@@ -160,11 +214,13 @@ function st2start(){
         fi
     done
 
-    # Register contents
-    echo 'Registering sensors, actions, rules, aliases, and policies...'
-    ./virtualenv/bin/python \
-        ./st2common/bin/st2-register-content \
-        --config-file $ST2_CONF --register-all
+    if [ "$load_content" = true ]; then
+        # Register contents
+        echo 'Registering sensors, actions, rules, aliases, and policies...'
+        ./virtualenv/bin/python \
+            ./st2common/bin/st2-register-content \
+            --config-file $ST2_CONF --register-all
+    fi
 
     # List screen sessions
     screen -ls || exit 0
@@ -183,7 +239,9 @@ function st2clean(){
     mongo st2 --eval "db.dropDatabase();"
     # start with clean logs
     LOGDIR=$(dirname $0)/../logs
-    rm ${LOGDIR}/*
+    if [ -d ${LOGDIR} ]; then
+        rm ${LOGDIR}/*
+    fi
     if [ -n "$ST2_EXPORTER" ]; then
         EXPORTS_DIR=$(exportsdir)
         echo "Removing $EXPORTS_DIR..."
@@ -191,7 +249,7 @@ function st2clean(){
     fi
 }
 
-case ${1} in
+case ${subcommand} in
 start)
     init
     st2start
@@ -211,6 +269,6 @@ restart)
     st2start
     ;;
 *)
-    echo "Usage: $0 [start|stop|restart|startclean]" >&2
+    usage
     ;;
 esac
